requirements.txt
pdfminer.six
docx2txt
nltk
customtkinter
packaging
pillow
